+++ changelog.md
@@ -5,11 +5,14 @@
 
 ## New features
 
+* Support fail_if_missing parameter for `gds.v2.model.delete`
+
 
 ## Bug fixes
 
 * Fixed a bug where `gds.v2.graph.list()` would fail due to `memoryUsage` not being present in the result.
 * Fixed a bug where GDS would fail if constructed via `from_driver` and without explicit auth provided.
+* Fixed a bug in `gds.v2.model.drop` where it would fail even if `fail_if_missing` was set to `False`. 
 
 
 ## Improvements